Inventory Sprite Changer
For now, when we pick up an item, the game will have an image on the first slot of the inventory to change into a pumpkin. It only applies to the pumpkin, sprite renderer activates when that slot is not empty.

Eventually, this will change into whatever sprite is assigned to a GameObject prefab/asset, or if we wanna make it more high quality, how the item should be viewed in the inventory and not out in the world.

We can always add more slots to the inventory, if need be.