PlantFetcher
Made a PlantFetcher.cs. The plan is to replace the plant objects (plantGrown, plantObj) within Planting.cs for the required plant.

Which will mean that PumpkinPlant.cs is essentially a plant which only contains data for the plant. I will also need a Harvester class of some sorts to sort plants into the inventory, along with their data.